pyproject.toml
[tool.poetry]
name = "abyssinian"
version = "0.1.0"
description = "Test project for deploying FastAPI app to AWS Lambda"
authors = ["oswalth <oswalth2@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
uvicorn = "^0.18.1"
psycopg2 = "^2.9.3"
pydantic = "^1.9.1"
FastAPI-SQLAlchemy = "^0.2.1"
environs = "^9.5.0"
alembic = "^1.8.0"
python-jose = {version = "^3.3.0", extras = ["cryptography"]}
passlib = {version = "^1.7.4", extras = ["bcrpyt"]}
python-multipart = "^0.0.5"
bcrypt = "^3.2.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
